# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce API",
    "version": "1.0.0",
    "description": "API documentation for Users, Products, Carts, and Orders"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
...
  "paths": {
...
    "/products": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a new product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "admin"
        ],
        "description": "Upload an image (multipart/form-data) and add product details",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "iPhone 16"
                  },
                  "description": {
                    "type": "string",
                    "example": "Latest iPhone model"
                  },
                  "category": {
                    "type": "string",
                    "example": "Mobiles"
                  },
                  "price": {
                    "type": "number",
                    "example": 799
                  },
                  "stock": {
                    "type": "number",
                    "example": 10
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "name",
                  "price",
                  "stock",
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          }
        }
      },
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products",
        "security": [],
        "x-roles": [],
        "responses": {
          "200": {
            "description": "List of products"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get single product by ID",
        "security": [],
        "x-roles": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "66d0e6f7b22a3c7c9d901234"
          }
        ],
        "responses": {
          "200": {
            "description": "Product details"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}